requirements.txt
pandas
pyarrow
tqdm
tensorboard
matplotlib
torcheval
torchinfo
# pip3 install torch --index-url https://download.pytorch.org/whl/cu118